Refactor deploy controllers to use RetryController
Followup to #1091 in support of #824.

Summary of the changes:

* Each controller is now in its own package (`pkg/deploy/controller/$type`) with a normalized name (`controller.go`, `factory.go`, etc.)
* `DeploymentController` has been split into `DeploymentController` and `DeployerPodController`
* All controllers now propagate semantic errors (e.g. fatal or nonfatal)
* Improved private interface usage throughout the controllers
* All controllers are now wrapped in a `RetryController` via their corresponding factories
