--- pkg/deploy/controller/factory/doc.go
@@ -1,2 +1,0 @@
-// Package factory contains code used to create deployment controllers.
-package factory
