+++ pkg/deploy/controller/imagechange/factory.go
@@ -1,0 +1,83 @@
+package imagechange
+
+import (
+	kapi "github.com/GoogleCloudPlatform/kubernetes/pkg/api"
+	"github.com/GoogleCloudPlatform/kubernetes/pkg/client/cache"
+	"github.com/GoogleCloudPlatform/kubernetes/pkg/labels"
+	"github.com/GoogleCloudPlatform/kubernetes/pkg/runtime"
+	kutil "github.com/GoogleCloudPlatform/kubernetes/pkg/util"
+	"github.com/GoogleCloudPlatform/kubernetes/pkg/watch"
+
+	osclient "github.com/openshift/origin/pkg/client"
+	controller "github.com/openshift/origin/pkg/controller"
+	deployapi "github.com/openshift/origin/pkg/deploy/api"
+	deployutil "github.com/openshift/origin/pkg/deploy/util"
+	imageapi "github.com/openshift/origin/pkg/image/api"
+)
+
+// ImageChangeControllerFactory can create an ImageChangeController which
+// watches all ImageRepository changes.
+type ImageChangeControllerFactory struct {
+	// Client is an OpenShift client.
+	Client osclient.Interface
+}
+
+// Create creates an ImageChangeController.
+func (factory *ImageChangeControllerFactory) Create() controller.RunnableController {
+	imageRepositoryLW := &deployutil.ListWatcherImpl{
+		ListFunc: func() (runtime.Object, error) {
+			return factory.Client.ImageRepositories(kapi.NamespaceAll).List(labels.Everything(), labels.Everything())
+		},
+		WatchFunc: func(resourceVersion string) (watch.Interface, error) {
+			return factory.Client.ImageRepositories(kapi.NamespaceAll).Watch(labels.Everything(), labels.Everything(), resourceVersion)
+		},
+	}
+	queue := cache.NewFIFO(cache.MetaNamespaceKeyFunc)
+	cache.NewReflector(imageRepositoryLW, &imageapi.ImageRepository{}, queue).Run()
+
+	deploymentConfigLW := &deployutil.ListWatcherImpl{
+		ListFunc: func() (runtime.Object, error) {
+			return factory.Client.DeploymentConfigs(kapi.NamespaceAll).List(labels.Everything(), labels.Everything())
+		},
+		WatchFunc: func(resourceVersion string) (watch.Interface, error) {
+			return factory.Client.DeploymentConfigs(kapi.NamespaceAll).Watch(labels.Everything(), labels.Everything(), resourceVersion)
+		},
+	}
+	store := cache.NewStore(cache.MetaNamespaceKeyFunc)
+	cache.NewReflector(deploymentConfigLW, &deployapi.DeploymentConfig{}, store).Run()
+
+	changeController := &ImageChangeController{
+		deploymentConfigClient: &deploymentConfigClientImpl{
+			listDeploymentConfigsFunc: func() ([]*deployapi.DeploymentConfig, error) {
+				configs := []*deployapi.DeploymentConfig{}
+				objs := store.List()
+				for _, obj := range objs {
+					configs = append(configs, obj.(*deployapi.DeploymentConfig))
+				}
+				return configs, nil
+			},
+			generateDeploymentConfigFunc: func(namespace, name string) (*deployapi.DeploymentConfig, error) {
+				return factory.Client.DeploymentConfigs(namespace).Generate(name)
+			},
+			updateDeploymentConfigFunc: func(namespace string, config *deployapi.DeploymentConfig) (*deployapi.DeploymentConfig, error) {
+				return factory.Client.DeploymentConfigs(namespace).Update(config)
+			},
+		},
+	}
+
+	return &controller.RetryController{
+		Queue:        queue,
+		RetryManager: controller.NewQueueRetryManager(queue, cache.MetaNamespaceKeyFunc, 1),
+		ShouldRetry: func(obj interface{}, err error) bool {
+			if _, isFatal := err.(fatalError); isFatal {
+				kutil.HandleError(err)
+				return false
+			}
+			return true
+		},
+		Handle: func(obj interface{}) error {
+			repo := obj.(*imageapi.ImageRepository)
+			return changeController.Handle(repo)
+		},
+	}
+}
